+++ src/utils/index.ts
@@ -1,1 +1,1 @@
-export * from './email-validator'
+export * from './email-validator-adapter'
